+++ crates/mimic/benches/data_key.rs
@@ -11,24 +11,18 @@ benchmark_group!(
 // create_data_key
 fn create_data_key(bench: &mut Bencher) {
     let key = "field_abc";
-    let value = Some(IndexValue::Text("value_xyz".to_string()));
+    let value = Some(IndexValue::Int(345345));
 
     bench.iter(|| {
-        let data_key = DataKey::new(vec![(key, value.clone())]);
+        let data_key = DataKey::new(vec![(key, value)]);
         std::hint::black_box(data_key);
     });
 }
 
 // compare_data_key
 fn compare_data_key(bench: &mut Bencher) {
-    let a = DataKey::new(vec![(
-        "field_abc",
-        Some(IndexValue::Text("value_xyz".to_string())),
-    )]);
-    let b = DataKey::new(vec![(
-        "field_abc",
-        Some(IndexValue::Text("value_xyz".to_string())),
-    )]);
+    let a = DataKey::new(vec![("field_abc", Some(IndexValue::Int(3)))]);
+    let b = DataKey::new(vec![("field_abc", Some(IndexValue::Int(4)))]);
 
     bench.iter(|| {
         let result = std::cmp::PartialEq::eq(&a, &b);
@@ -38,10 +32,8 @@ fn compare_data_key(bench: &mut Bencher) {
 
 // compare_vec_value
 fn compare_vec_value(bench: &mut Bencher) {
-    let a: Vec<(String, Option<IndexValue>)> = vec![(
-        "field_abc".to_string(),
-        Some(IndexValue::Text("value_xyz".to_string())),
-    )];
+    let a: Vec<(String, Option<IndexValue>)> =
+        vec![("field_abc".to_string(), Some(IndexValue::Int(123123)))];
     let b = a.clone();
 
     bench.iter(|| {