+++ crates/mimic_tests/design/src/admin/mod.rs
@@ -1,6 +1,11 @@
 use crate::prelude::*;
 
 ///
+/// ADMIN TESTS
+/// set up to test the admin interface
+///
+
+///
 /// ComplexEntity
 ///