+++ crates/mimic_tests/design/src/fixture/mod.rs
@@ -1,0 +1,6 @@
+pub mod rarity;
+
+//
+// FIXTURE TESTS
+// for testing Endpoints, relations etc.
+//