+++ crates/mimic_tests/src/lib.rs
@@ -1,0 +1,6 @@
+//
+// MIMIC TESTS
+//
+
+#[cfg(test)]
+pub mod test {}