+++ crates/test/design/src/constant.rs
@@ -1,8 +1,0 @@
-use crate::prelude::*;
-
-///
-/// Stringy
-///
-
-#[constant(ty = "Str", value = "Hello")]
-pub struct STRINGY {}